chore: Fix circular dependency in tanstack-table.d.ts
